Move Replacer and ParentReplacer into src/Replace/
