+++ docs/architecture.md
@@ -145,6 +145,8 @@ src/
   Config/                           # Configuration models (Mozart, Package, Psr4, Classmap, Files, etc.)
   Composer/Autoload/                # Autoloader abstractions (NamespaceAutoloader, AbstractAutoloader)
   Replace/
+    Replacer.php                    # Orchestrator (routes packages to correct replacer)
+    ParentReplacer.php              # Cross-replacement: propagates renames into parent packages
     AbstractAutoloadReplacer.php    # Abstract base (holds autoloader reference)
     AutoloadReplacer.php            # Interface: extends StringReplacer with setAutoloader()
     StringReplacer.php              # Interface: same as Replacer but for class-map style
@@ -152,8 +154,6 @@ src/
     Namespace/                      # NamespaceReplacer, PrefixVisitor
     Support/                        # AstUtils, ExistenceCheckTrait, NameNodeContextTrait
   Mover.php                         # Copies files from vendor/ to target directories
-  Replacer.php                      # Orchestrator (routes packages to correct replacer)
-  ParentReplacer.php                # Cross-replacement: propagates renames into parent packages
   PackageFinder.php                 # Dependency tree resolution (BFS with deduplication)
   PackageFactory.php                # Creates Package objects from composer.json
   FilesHandler.php                  # File I/O via Flysystem (read, write, copy, delete)
