+++ tests/Unit/Replace/ReplacerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CoenJacobs\Mozart\Tests\Unit;
+namespace CoenJacobs\Mozart\Tests\Unit\Replace;
 
 use CoenJacobs\Mozart\Config\Classmap;
 use CoenJacobs\Mozart\Config\Mozart;
@@ -10,7 +10,7 @@ use CoenJacobs\Mozart\Config\Package;
 use CoenJacobs\Mozart\Config\Psr4;
 use CoenJacobs\Mozart\Replace\Classmap\ClassmapReplacer;
 use CoenJacobs\Mozart\Replace\Namespace\NamespaceReplacer;
-use CoenJacobs\Mozart\Replacer;
+use CoenJacobs\Mozart\Replace\Replacer;
 use PHPUnit\Framework\TestCase;
 use stdClass;
